Reworked storage feature to not concat files, replaced blocking std File with async tokio File; Implemented AES functionality to cryptfns crate and exposed them through wasm to be available in the frontend instead of using js implementations

+++ storage/src/data/create_file.rs
@@ -35,8 +35,8 @@ pub struct CreateFile {
     /// Mime type of the file or "dir" for directory
     pub mime: Option<String>,
     /// Total size of the file
-    pub size: Option<i64>,
-    /// Total number of chunks, must be ceil(size / CHUNK_SIZE_BYTES)
+    pub size: Option<i32>,
+    /// Total number of chunks, no limitations, frontend can decide
     pub chunks: Option<i32>,
     /// ID of the directory the file is located in (directories are files too)
     pub file_id: Option<i32>,
